feat: add role and user management components with CRUD functionality - Implemented Role Management with create, edit, and delete capabilities. - Added User Management to assign roles to users. - Introduced services for handling roles and users. - Enhanced permission handling with bitmasking for roles. - Created UI components for roles and users with search and loading states. - Created OfficeService for managing office data. - Defined types for Role, User, and Office.

+++ smart-clicker-frontend/src/app/pages/qotd/qotd.component.ts
@@ -93,7 +93,7 @@ export class QotdComponent implements OnInit {
   }
 
   async getQotds() {
-    await this.qotdService.getQotdsForOffice("Office_A1").then((qotds) => {
+    await this.qotdService.getQotdsForOffice("f17120a9-4d72-402a-812e-6cd2b7482d6a").then((qotds) => {
       this.qotds = qotds.map((qotd) => {
         return {
           id: qotd.id,
@@ -149,6 +149,10 @@ export class QotdComponent implements OnInit {
 
   async deleteEvent(eventToDelete: CalendarEvent): Promise<void> {
     try {
+      if (eventToDelete.start < new Date()) {
+        this.toastService.error("You cannot delete a QOTD in the past.");
+        return;
+      }
       await this.qotdService.deleteQotd(
         eventToDelete.id as string,
         eventToDelete.meta.officeSpaceId
@@ -254,7 +258,7 @@ export class QotdComponent implements OnInit {
       id: crypto.randomUUID(),
       question: this.addForm.value.question,
       date: new Date(this.addForm.value.date),
-      officeSpaceId: "Office_A1", // or get from context
+      officeSpaceId: "f17120a9-4d72-402a-812e-6cd2b7482d6a",
     };
     try {
       const response = await this.qotdService.createQotd(newQotd);
